requirement.txt
#langchain core
langchain
langchain-core

#openAi integration
langchain-openai
openAi

#Anthropic integration
langchain-anthropic  

#Google Gemini (PaLM) integration

langchain-google-genai
google-generativeai

#hugging face integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

#Machine Learning Utilities
numpy 
scikit-learn